[ refactor ] DTO 이름 변경 ( #16 )
