+++ src/main/java/com/qring/queue/application/v2/service/QueueServiceV2.java
@@ -2,7 +2,7 @@ package com.qring.queue.application.v2.service;
 
 import com.qring.queue.application.v2.message.redis.RedisMessagePublisherV2;
 import com.qring.queue.application.v2.res.QueueGetResDTOV2;
-import com.qring.queue.infrastructure.messaging.dto.ReservationEventDTOV2;
+import com.qring.queue.infrastructure.messaging.dto.ReservationCreationEventDTOV2;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,7 +17,7 @@ public class QueueServiceV2 {
     private final RedisMessagePublisherV2 redisMessagePublisherV2;
 
     // 대기열 등록
-    public void enrollWaitingListByEvent(ReservationEventDTOV2 event) {
+    public void enrollWaitingListByEvent(ReservationCreationEventDTOV2 event) {
 
         // Redis에 예약 정보 저장
         redisMessagePublisherV2.addWaitingListByRestaurantIdAndReservationId(
@@ -29,7 +29,7 @@ public class QueueServiceV2 {
                 event.getReservationInfo().getId());
     }
 
-    // 사용자 대기 순서와 총 대기 인원 반환
+    // 사용자 대기 순서 반환
     public QueueGetResDTOV2 getBy(Long restaurantId, Long reservationId) {
 
         return redisMessagePublisherV2.getQueueInfoByRestaurantIdAndReservationId(restaurantId, reservationId);
